+++ sat/cli/showrev/parser.py
@@ -45,8 +45,8 @@ def add_showrev_subparser(subparsers):
 
     showrev_parser.add_argument(
         '--all',
-        help='Display everything. Equivalent to specifying --system, --products, '
-             '--docker, --packages, and --release-files.',
+        help='Display everything. Equivalent to specifying --system, --products,'
+             '--local and --release-files',
         action='store_true')
     showrev_parser.add_argument(
         '--system',
@@ -59,15 +59,15 @@ def add_showrev_subparser(subparsers):
              'This is enabled when no other options are specified.',
         action='store_true')
     showrev_parser.add_argument(
+        '--local',
+        help='Display system version information specific to the host on which '
+             'the command is running. This is enabled when no other options '
+             'are specified.',
+        action='store_true')
+    showrev_parser.add_argument(
         '--release-files',
         help='Display version information about locally installed release files.',
         action='store_true')
-    showrev_parser.add_argument(
-        '--docker',
-        help='Display running docker image versions.',
-        action='store_true')
-    showrev_parser.add_argument(
-        '--packages', help='Display installed rpm versions.', action='store_true')
     showrev_parser.add_argument(
         '--sitefile',
         help='Specify custom site information file printed by --system.',