+++ tests/cli/showrev/test_system.py
@@ -22,7 +22,6 @@ ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 OTHER DEALINGS IN THE SOFTWARE.
 """
 
-import codecs
 import os
 import subprocess
 import unittest
@@ -37,66 +36,10 @@ import sat.cli.showrev.system
 samples = os.path.join(os.path.dirname(__file__), 'samples')
 
 
-# Helpers for mock-decorations.
-
-def bad_zypper_return(package):
-    """Used by mock to replace sat.cli.showrev.system.subprocess.check_output.
-
-    Used to send 'unreadable' output to within zypper_seach method, which
-    commands zypper to report in xml.
-
-    Args:
-        package: Used to insert a 'real' package name into the result.
-    """
-    return codecs.encode(' | {} | package | 19.05.0-6 | x86_64 | Cray Module SLES'.format(package))
-
-
-def zypper_good_xml(packages):
-    """Reads from zypper-good-xml to pretend the zypper query succeeded.
-
-    Args:
-        package: Used to format the string inside the file to insert the
-            package name.
-
-    Returns:
-        The contents of the file formatted with the package name encoded
-        as a byte string. This is the expected return format of
-        subprocess.check_output .
-    """
-    s = ''
-    with open('{}/zypper-good-xml'.format(samples), 'r') as f:
-        s = f.read().format(packages[-1])
-    return codecs.encode(s)
-
-
 class TestSystem(unittest.TestCase):
 
     def tearDown(self):
         mock.patch.stopall()
-
-    @mock.patch('sat.cli.showrev.system.subprocess.check_output', bad_zypper_return)
-    def test_get_zypper_versions_bad_return(self):
-        """Ensures the get_zypper_versions method return 'ERROR' on bad output.
-        """
-        result = sat.cli.showrev.system.get_zypper_versions(['slurm-slurmd'])
-        self.assertEqual(result['slurm-slurmd'], 'ERROR')
-
-    @mock.patch('sat.cli.showrev.system.subprocess.check_output', zypper_good_xml)
-    def test_get_zypper_versions_good_return(self):
-        """Ensures the get_zypper_versions method can parse correct xml return.
-        """
-        result = sat.cli.showrev.system.get_zypper_versions(['slurm-slurmd'])
-        self.assertEqual(result['slurm-slurmd'], '19.05.0-6')
-
-    @mock.patch(
-        'sat.cli.showrev.system.subprocess.check_output',
-        side_effect=sat.cli.showrev.system.subprocess.CalledProcessError(
-            cmd=['echo'], returncode=104))
-    def test_get_zypper_versions_package_not_found(self, mocksubprocess):
-        """The get_zypper_versions method should return None if zypper query failed.
-        """
-        result = sat.cli.showrev.system.get_zypper_versions(['idontexist'])
-        self.assertIs(result['idontexist'], None)
 
     @mock.patch(
         'sat.cli.showrev.system._get_hsm_components',
@@ -120,88 +63,6 @@ class TestSystem(unittest.TestCase):
         """
         result = sat.cli.showrev.system.get_interconnects()
         self.assertEqual(['ERROR'], result)
-
-    def test_get_sles_version_correct_format(self):
-        """Positive test case for get_sles_version.
-        """
-        file_ = '{}/os-release'.format(samples)
-        with open(file_) as f:
-            data = f.read()
-
-        mock_open = mock.mock_open(read_data=data)
-
-        with mock.patch('sat.cli.showrev.system.open', mock_open):
-            short = '{}/os-release'.format(samples)
-            result = sat.cli.showrev.system.get_sles_version()
-            self.assertTrue(os.path.exists(short))
-            self.assertEqual(result, 'SLES 15')
-
-    def test_get_sles_version_file_not_found(self):
-        """get_sles_version should return a certain error string on file-not-found.
-
-        The underlying path to os-release is hardcoded in this function, so its
-        error string won't match the path in our lambda. That's expected.
-        """
-        mock_open = mock.mock_open()
-        mock_open.side_effect = FileNotFoundError
-
-        with mock.patch('sat.cli.showrev.system.open', mock_open):
-            result = sat.cli.showrev.system.get_sles_version()
-            self.assertEqual(result, 'ERROR')
-            self.assertFalse(os.path.exists('idontexist'))
-
-    def test_get_sles_version_empty_osrel(self):
-        """get_sles_version should return error message on empty read.
-        """
-        file_ = '{}/empty'.format(samples)
-        with open(file_) as f:
-            data = f.read()
-
-        mock_open = mock.mock_open(read_data=data)
-
-        with mock.patch('sat.cli.showrev.system.open', mock_open):
-            short = '{}/empty'.format(samples)
-            result = sat.cli.showrev.system.get_sles_version()
-            self.assertTrue(os.path.exists(short))
-            self.assertEqual(result, 'ERROR')
-
-    def test_get_sles_version_missing_sles(self):
-        """Test behavior if /etc/os-release is missing required fields.
-        """
-        file_ = '{}/os-release-missing-sles'.format(samples)
-        with open(file_) as f:
-            data = f.read()
-
-        mock_open = mock.mock_open(read_data=data)
-
-        with mock.patch('sat.cli.showrev.system.open', mock_open):
-            short = '{}/os-release-missing-sles'.format(samples)
-            result = sat.cli.showrev.system.get_sles_version()
-            self.assertTrue(os.path.exists(short))
-            self.assertEqual(result, 'ERROR')
-
-    def test_get_sles_version_bad_osrel_permissions(self):
-        """get_sles_version should report an error on bad permissions.
-        """
-        mock_open = mock.mock_open()
-        mock_open.side_effect = PermissionError
-
-        with mock.patch('sat.cli.showrev.system.open', mock_open):
-            result = sat.cli.showrev.system.get_sles_version()
-            self.assertEqual(result, 'ERROR')
-
-    def test_get_sles_version_name_and_version_empty(self):
-        """get_sles_version should report error on empty NAME or VERSION fields.
-        """
-        file_ = '{}/os-release-empty-sles'.format(samples)
-        with open(file_) as f:
-            data = f.read()
-
-        mock_open = mock.mock_open(read_data=data)
-
-        with mock.patch('sat.cli.showrev.system.open', mock_open):
-            result = sat.cli.showrev.system.get_sles_version()
-            self.assertEqual(result, 'ERROR')
 
     def test_get_slurm_version_config_exception(self):
         """get_slurm_version kubernetes config failure test.